refactor(tests): move set_agency, set_verifier to conftest
make these helpers available to other test suites
